Refactoring enrollment system
