+++ src/main/java/ro/kepler/rominfo/beans/EnrollmentView.java
@@ -1,0 +1,82 @@
+package ro.kepler.rominfo.beans;
+
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
+import org.primefaces.context.RequestContext;
+import ro.kepler.rominfo.dto.CourseDto;
+import ro.kepler.rominfo.model.Course;
+import ro.kepler.rominfo.model.Student;
+import ro.kepler.rominfo.service.CourseService;
+import ro.kepler.rominfo.service.EnrollmentService;
+import ro.kepler.rominfo.service.StudentService;
+
+import javax.faces.application.FacesMessage;
+import javax.faces.bean.ManagedBean;
+import javax.faces.bean.ManagedProperty;
+import javax.faces.bean.ViewScoped;
+import javax.faces.context.FacesContext;
+
+/**
+ * Created by Dragos on 22.05.2017.
+ */
+
+@ManagedBean
+@ViewScoped
+public class EnrollmentView {
+
+    @ManagedProperty("#{studentService}")
+    private StudentService studentService;
+
+    @ManagedProperty("#{courseService}")
+    private CourseService courseService;
+
+    @ManagedProperty("#{enrollmentService}")
+    private EnrollmentService enrollmentService;
+
+    public void setCourseService(CourseService courseService) {
+        this.courseService = courseService;
+    }
+
+
+
+    public void setEnrollmentService(EnrollmentService enrollmentService) {
+        this.enrollmentService = enrollmentService;
+    }
+
+    private static final Log LOGGER = LogFactory.getLog(EnrollmentView.class);
+
+
+    public void setStudentService(StudentService studentService) {
+        this.studentService = studentService;
+    }
+
+    public String enroll(String email, CourseDto courseDto) {
+        Student student = studentService.getStudentByEmail(email);
+        Course course = courseService.getCourseByName(courseDto.getCourseName());
+
+        try {
+            if(!enrollmentService.alreadyEnrolled(student, course)) {
+                enrollmentService.enroll(student, course);
+                LOGGER.info("enrollment successful for " + student.getEmail());
+                RequestContext.getCurrentInstance().showMessageInDialog(new FacesMessage(FacesMessage.SEVERITY_INFO, "Enrollment Successful", "Now you have enrolled to " + courseDto.getCourseName() + " course!"));
+                RequestContext.getCurrentInstance().execute("PF('courseDialog').hide();");
+                return "success";
+            }
+            else {
+                LOGGER.info("student already enrolled for this course");
+                RequestContext.getCurrentInstance().showMessageInDialog(new FacesMessage(FacesMessage.SEVERITY_ERROR, "Enrollment Failed", "You are already enrolled on that course!"));
+                RequestContext.getCurrentInstance().execute("PF('courseDialog').hide();");
+                return "failed";
+            }
+
+        } catch (Exception ex) {
+            ex.printStackTrace();
+            LOGGER.info("enrollment failed for " + email);
+            FacesContext.getCurrentInstance().addMessage(null,
+                    new FacesMessage(FacesMessage.SEVERITY_WARN,
+                            "Enrollment failed.",
+                            "The enrollment process has failed."));
+            return null;
+        }
+    }
+}
