+++ src/main/java/ro/kepler/rominfo/dto/CourseDto.java
@@ -1,0 +1,41 @@
+package ro.kepler.rominfo.dto;
+
+import javax.faces.bean.ManagedBean;
+import javax.faces.bean.RequestScoped;
+import java.io.Serializable;
+
+/**
+ * Created by Dragos on 19.05.2017.
+ */
+
+@ManagedBean
+@RequestScoped
+public class CourseDto implements Serializable {
+    private String courseName;
+    private String category;
+    private String professor;
+
+    public String getCourseName() {
+        return courseName;
+    }
+
+    public void setCourseName(String courseName) {
+        this.courseName = courseName;
+    }
+
+    public String getCategory() {
+        return category;
+    }
+
+    public void setCategory(String category) {
+        this.category = category;
+    }
+
+    public String getProfessor() {
+        return professor;
+    }
+
+    public void setProfessor(String professor) {
+        this.professor = professor;
+    }
+}
