+++ src/main/java/ro/kepler/rominfo/model/Professor.java
@@ -1,0 +1,67 @@
+package ro.kepler.rominfo.model;
+
+import java.io.Serializable;
+
+/**
+ * Created by Dragos on 19.05.2017.
+ */
+public class Professor implements Serializable {
+    private int professorId;
+    private String firstName;
+    private String lastName;
+    private long ssn;
+    private String email;
+    private String password;
+
+    public Integer getProfessorId() {
+        return professorId;
+    }
+
+    public void setProfessorId(Integer professorId) {
+        this.professorId = professorId;
+    }
+
+    public String getFirstName() {
+        return firstName;
+    }
+
+    public void setFirstName(String firstName) {
+        this.firstName = firstName;
+    }
+
+    public String getLastName() {
+        return lastName;
+    }
+
+    public void setLastName(String lastName) {
+        this.lastName = lastName;
+    }
+
+    public long getSsn() {
+        return ssn;
+    }
+
+    public void setSsn(long ssn) {
+        this.ssn = ssn;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+
+    public String getFullName() {
+        return firstName + " " + lastName;
+    }
+}
