--- src/main/resources/maps/CourseMapper.xml
@@ -1,42 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-
-<mapper namespace="ro.kepler.rominfo.mapper.CourseMapper">
-
-    <resultMap type="Course" id="CourseResult">
-        <id property="courseCode" column="coursecode"/>
-        <result property="courseName" column="coursename"/>
-        <result property="category" column="category"/>
-        <association property="professor" resultMap="ProfessorResult" columnPrefix="prof_"/>
-    </resultMap>
-
-    <resultMap type="Professor" id="ProfessorResult">
-        <id property="professorId" column="professorId"/>
-        <result property="firstName" column="firstname"/>
-        <result property="lastName" column="lastname"/>
-    </resultMap>
-
-    <select id="getCourseByCode" parameterType="int" resultMap="CourseResult">
-        SELECT
-        c."coursecode" as courseCode, c."coursename" AS courseName, c."category" AS category, c."professorid" AS professorId
-        FROM public."course" c
-        WHERE c."coursecode"=#{courseCode}
-    </select>
-
-    <select id="getCourseByName" parameterType="java.lang.String" resultMap="CourseResult">
-        SELECT
-        c."coursecode" as courseCode, c."coursename" AS courseName, c."category" AS category
-        FROM public."course" c
-        WHERE c."coursename"=#{courseName}
-    </select>
-
-    <select id="getAllCourses" resultMap="CourseResult">
-        SELECT
-        c."coursecode" as courseCode, c."coursename" AS courseName, c."category" AS category,
-        p."firstname" as prof_firstname, p."lastname" as prof_lastname, p."professorid" as prof_professorid
-        FROM public."course" c INNER JOIN professor p
-        ON p.professorId = c.professorId
-    </select>
-
-</mapper>
