--- src/main/resources/maps/EnrollmentMapper.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-
-<mapper namespace="ro.kepler.rominfo.mapper.EnrollmentMapper">
-
-    <resultMap id="StudentResult" type="ro.kepler.rominfo.model.Course">
-        <id column="coursecode" property="courseCode"/>
-        <result column="coursename" property="courseName"/>
-        <result column="category" property="category"/>
-        <collection property="students" resultMap="CourseResult" />
-    </resultMap>
-
-    <resultMap id="CourseResult" type="ro.kepler.rominfo.model.Student">
-        <id column="studentid" property="studentId"/>
-        <result column="firstname"  property="firstName"/>
-        <result column="lastname" property="lastName"/>
-        <result column="ssn" property="ssn"/>
-        <result column="email"  property="email"/>
-        <result column="password" property="password"/>
-        <collection property="courses" resultMap="StudentResult" />
-    </resultMap>
-
-    <insert id="addEnrollment">
-        INSERT INTO enrollment (studentid, coursecode)
-        values (#{studentId},#{courseCode})
-    </insert>
-
-</mapper>
