+++ src/main/resources/mybatis-config.xml
@@ -1,0 +1,13 @@
+<?xml version='1.0' encoding='UTF-8' ?>
+<!DOCTYPE configuration	PUBLIC '-//mybatis.org//DTD Config 3.0//EN'	'http://mybatis.org/dtd/mybatis-3-config.dtd'>
+<configuration>
+    <settings>
+        <setting name="useGeneratedKeys" value="true"/>
+        <setting name="jdbcTypeForNull" value="NULL"/>
+    </settings>
+
+    <typeAliases>
+        <package name="ro.kepler.rominfo.model"/>
+    </typeAliases>
+
+</configuration>
