+++ src/main/resources/ro/kepler/rominfo/mappers/EnrollmentMapper.xml
@@ -1,0 +1,47 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+
+<mapper namespace="ro.kepler.rominfo.mappers.EnrollmentMapper">
+
+    <resultMap id="StudentResult" type="ro.kepler.rominfo.model.Course">
+        <id column="coursecode" property="courseCode"/>
+        <result column="coursename" property="courseName"/>
+        <result column="category" property="category"/>
+        <collection property="students" resultMap="CourseResult" />
+    </resultMap>
+
+    <resultMap id="CourseResult" type="ro.kepler.rominfo.model.Student">
+        <id column="studentid" property="studentId"/>
+        <result column="firstname"  property="firstName"/>
+        <result column="lastname" property="lastName"/>
+        <result column="ssn" property="ssn"/>
+        <result column="email"  property="email"/>
+        <result column="password" property="password"/>
+        <collection property="courses" resultMap="StudentResult" />
+    </resultMap>
+
+    <select id="getCoursesOfStudent" resultMap="StudentResult">
+        SELECT
+        c.coursecode AS courseCode,
+        c.coursename AS courseName,
+        c.category AS category
+        FROM course c
+        INNER JOIN enrollment e
+        ON c.coursecode = e.coursecode
+        INNER JOIN student s
+        ON e.studentid = s.studentid
+        WHERE s.studentid =#{studentId}
+    </select>
+
+    <insert id="addEnrollment">
+        INSERT INTO enrollment (studentid, coursecode)
+        values (#{studentId},#{courseCode})
+    </insert>
+
+    <select id="find">
+
+
+    </select>
+
+</mapper>
