+++ src/main/resources/ro/kepler/rominfo/mappers/StudentMapper.xml
@@ -1,0 +1,45 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+
+<mapper namespace="ro.kepler.rominfo.mappers.StudentMapper">
+
+    <resultMap type="student" id="StudentResult">
+        <id property="studentId" column="studentId"/>
+        <result property="firstName" column="firstName"/>
+        <result property="lastName" column="lastName"/>
+        <result property="ssn" column="ssn"/>
+        <result property="email" column="email"/>
+        <result property="password" column="password"/>
+    </resultMap>
+
+    <select id="getStudentById" parameterType="int" resultMap="StudentResult">
+        SELECT
+        s."studentid" as studentid, s."firstname" AS firstName, s."lastname" AS lastName,
+        s."ssn" AS ssn, s."email" AS email, s."password" AS password
+        FROM public."student" s
+        WHERE s."studentId"=#{studentId}
+    </select>
+
+    <select id="getAllStudents" resultMap="StudentResult">
+        SELECT
+        s."studentid" as studentId, s."firstname" AS firstName, s."lastname" AS lastName,
+        s."ssn" AS ssn, s."email" AS email, s."password" AS password
+        FROM public."student" s
+    </select>
+
+    <insert id="addStudent" useGeneratedKeys="true"
+            keyProperty="studentId">
+        insert into student (firstName,lastName,ssn,email,password)
+        values (#{firstName},#{lastName},#{ssn},#{email},#{password})
+    </insert>
+
+    <select id="findByEmail" resultMap="StudentResult">
+        SELECT
+            s."studentid" as studentId, s."firstname" AS firstName, s."lastname" AS lastName,
+            s."ssn" AS ssn, s."email" AS email, s."password" AS password
+        FROM public."student" s
+        WHERE s."email"=#{email}
+    </select>
+
+</mapper>
