--- src/main/ro/kepler/rominfo/dto/RegisterDto.java
@@ -1,58 +1,0 @@
-package ro.kepler.rominfo.dto;
-
-import ro.kepler.rominfo.service.StudentService;
-
-import java.io.Serializable;
-
-/**
- * Created by Dragos on 15.05.2017.
- */
-public class RegisterDto implements Serializable {
-
-    private String firstName;
-    private String lastName;
-    private long ssn;
-    private String email;
-    private String password;
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
-
-    public long getSsn() {
-        return ssn;
-    }
-
-    public void setSsn(long ssn) {
-        this.ssn = ssn;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-}
