--- src/main/ro/kepler/rominfo/model/Student.java
@@ -1,80 +1,0 @@
-package ro.kepler.rominfo.model;
-
-import java.io.Serializable;
-import java.util.ArrayList;
-import java.util.List;
-
-/**
- * Created by Dragos on 15.05.2017.
- */
-public class Student implements Serializable {
-    private Integer studentId;
-    private String firstName;
-    private String lastName;
-    private long ssn;
-    private String email;
-    private String password;
-
-    private List<Course> courses = new ArrayList<Course>();
-
-    public int getStudentId() {
-        return studentId;
-    }
-
-    public void setStudentId(Integer studentId) {
-        this.studentId = studentId;
-    }
-
-    public List<Course> getCourses() {
-        return courses;
-    }
-
-    public void setCourses(List<Course> courses) {
-        this.courses = courses;
-    }
-
-    public void setStudentId(int studentId) {
-        this.studentId = studentId;
-
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
-
-    public long getSsn() {
-        return ssn;
-    }
-
-    public void setSsn(long ssn) {
-        this.ssn = ssn;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-}
