--- src/main/ro/kepler/rominfo/service/CourseService.java
@@ -1,13 +1,0 @@
-package ro.kepler.rominfo.service;
-
-import ro.kepler.rominfo.model.Course;
-
-import java.util.List;
-
-/**
- * Created by Dragos on 19.05.2017.
- */
-public interface CourseService {
-    List<Course> getAllCourses();
-    Course find(String courseName);
-}
