--- src/main/ro/kepler/rominfo/service/EnrollmentService.java
@@ -1,11 +1,0 @@
-package ro.kepler.rominfo.service;
-
-import ro.kepler.rominfo.model.Course;
-import ro.kepler.rominfo.model.Student;
-
-/**
- * Created by Dragos on 22.05.2017.
- */
-public interface EnrollmentService {
-    void enroll(Student student, Course course);
-}
