+++ src/main/webapp/WEB-INF/applicationContext.xml
@@ -3,12 +3,12 @@
 <beans xmlns="http://www.springframework.org/schema/beans"
        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
        xmlns:context="http://www.springframework.org/schema/context"
-       xmlns:mvc="http://www.springframework.org/schema/mvc"
-       xmlns:tx="http://www.springframework.org/schema/tx"
-       xsi:schemaLocation="http://www.springframework.org/schema/mvc http://www.springframework.org/schema/mvc/spring-mvc-3.2.xsd
-      http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd
+       xmlns:tx="http://www.springframework.org/schema/tx" xmlns:mybatis="http://mybatis.org/schema/mybatis-spring"
+       xsi:schemaLocation= "http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd
       http://www.springframework.org/schema/tx http://www.springframework.org/schema/tx/spring-tx-3.1.xsd
-      http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context-3.1.xsd">
+      http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context-3.1.xsd http://mybatis.org/schema/mybatis-spring http://mybatis.org/schema/mybatis-spring.xsd">
+
+    <context:component-scan base-package="ro.kepler.rominfo" />
 
     <bean id="dataSource" class="org.apache.commons.dbcp.BasicDataSource">
         <property name="driverClassName" value="org.postgresql.Driver"/>
@@ -20,70 +20,15 @@
     <bean id="sqlSessionFactory" class="org.mybatis.spring.SqlSessionFactoryBean">
         <property name="dataSource" ref="dataSource" />
         <property name="typeAliasesPackage" value="ro.kepler.rominfo.model"/>
-        <property name="mapperLocations" value="classpath:maps/*.xml" />
+        <property name="mapperLocations" value="classpath:ro/kepler/rominfo/mappers/*.xml" />
     </bean>
-
-    <bean id="sqlSession" class="org.mybatis.spring.SqlSessionTemplate">
-        <constructor-arg index="0" ref="sqlSessionFactory" />
-    </bean>
-
-    <bean class="org.mybatis.spring.mapper.MapperScannerConfigurer">
-        <property name="basePackage" value="ro.kepler.rominfo.mapper" />
-    </bean>
-
-    <tx:annotation-driven transaction-manager="transactionManager"/>
 
     <bean id="transactionManager" class="org.springframework.jdbc.datasource.DataSourceTransactionManager">
         <property name="dataSource" ref="dataSource" />
     </bean>
 
-    <context:component-scan base-package="ro.kepler.rominfo.service" />
+    <tx:annotation-driven transaction-manager="transactionManager" order="1"/>
 
-    <bean id="studentMapper" class="org.mybatis.spring.mapper.MapperFactoryBean">
-        <property name="mapperInterface" value="ro.kepler.rominfo.mapper.StudentMapper" />
-        <property name="sqlSessionFactory" ref="sqlSessionFactory" />
-    </bean>
-
-    <bean id="courseMapper" class="org.mybatis.spring.mapper.MapperFactoryBean">
-        <property name="mapperInterface" value="ro.kepler.rominfo.mapper.CourseMapper" />
-        <property name="sqlSessionFactory" ref="sqlSessionFactory" />
-    </bean>
-
-    <bean id="enrollmentMapper" class="org.mybatis.spring.mapper.MapperFactoryBean">
-        <property name="mapperInterface" value="ro.kepler.rominfo.mapper.EnrollmentMapper" />
-        <property name="sqlSessionFactory" ref="sqlSessionFactory" />
-    </bean>
-
-    <bean id = "courseService" class = "ro.kepler.rominfo.service.impl.CourseServiceImpl">
-        <property name="courseMapper" ref="courseMapper" />
-    </bean>
-
-    <bean id = "studentService" class = "ro.kepler.rominfo.service.impl.StudentServiceImpl">
-        <property name="studentMapper" ref="studentMapper" />
-    </bean>
-
-    <bean id = "loginHandler" class = "ro.kepler.rominfo.controller.LoginHandler">
-        <property name="studentService" ref="studentService" />
-    </bean>
-
-    <bean id = "registerHandler" class = "ro.kepler.rominfo.controller.RegisterHandler">
-        <property name="studentService" ref="studentService" />
-    </bean>
-
-    <bean id = "courseHandler" class = "ro.kepler.rominfo.controller.CourseHandler">
-        <property name="courseService" ref="courseService" />
-    </bean>
-
-    <bean id = "enrollmentHandler" class = "ro.kepler.rominfo.controller.EnrollmentHandler">
-        <property name="studentService" ref="studentService" />
-        <property name="courseService" ref="courseService" />
-        <property name="enrollmentService" ref="enrollmentService" />
-    </bean>
-
-    <bean id="enrollmentService" class="ro.kepler.rominfo.service.impl.EnrollmentServiceImpl">
-        <property name="enrollmentMapper" ref="enrollmentMapper" />
-    </bean>
-
-
-
+    <!-- DAO BEANS AND MAPPERS-->
+    <mybatis:scan base-package="ro.kepler.rominfo.mappers" factory-ref="sqlSessionFactory"/>
 </beans>
