--- src/main/webapp/resources/css/common.css
@@ -1,12 +1,0 @@
-.ui-layout-unit.topbar {
-    padding: 0;
-}
-
-.ui-layout-unit.topbar > div {
-    background-color: black;
-    /*height: 40px;*/
-}
-
-.logo {
-    float: left;
-}
