+++ src/main/webapp/secured/viewAllCourses.xhtml
@@ -19,15 +19,15 @@
         <p:panel header="Home">
             <p>
                 <p:outputLabel
-                        value="Welcome, #{loginHandler.currentStudent.firstName}!"/>
+                        value="Welcome, #{loginView.email}!"/>
                 <p:commandButton style="float:right" value="Logout"
-                                 action="#{loginHandler.logout()}"/>
+                                 action="#{loginView.logout()}"/>
             </p>
         </p:panel>
         <p:growl id="msgs" showDetail="true"/>
 
         <br/><br/>
-        <p:dataTable value="#{courseHandler.getAllCourses()}" var="courseDto">
+        <p:dataTable value="#{courseView.getAllCourses()}" var="courseDto">
             <f:facet name="header">
                 List of Courses
             </f:facet>
@@ -48,31 +48,31 @@
             <p:column style="width:32px;text-align: center">
                 <p:commandButton update=":form:courseDetail" oncomplete="PF('courseDialog').show()"
                                  icon="ui-icon-search" title="View">
-                    <f:setPropertyActionListener value="#{courseDto}" target="#{courseSelection.selectedCourse}"/>
+                    <f:setPropertyActionListener value="#{courseDto}" target="#{courseView.selectedCourse}"/>
                 </p:commandButton>
             </p:column>
         </p:dataTable>
 
 
-    <p:dialog header="Course Info" widgetVar="courseDialog" modal="true" showEffect="fade" hideEffect="fade"
+   <p:dialog header="Course Info" widgetVar="courseDialog" modal="true" showEffect="fade" hideEffect="fade"
               resizable="false">
         <p:outputPanel id="courseDetail" style="text-align:center;">
-            <p:panelGrid columns="2" rendered="#{not empty courseSelection.selectedCourse}" columnClasses="label,value">
+            <p:panelGrid columns="2" rendered="#{not empty courseView.selectedCourse}" columnClasses="label,value">
 
                 <h:outputText value="Name:"/>
-                <h:outputText value="#{courseSelection.selectedCourse.courseName}"/>
+                <h:outputText value="#{courseView.selectedCourse.courseName}"/>
 
                 <h:outputText value="Category"/>
-                <h:outputText value="#{courseSelection.selectedCourse.category}"/>
+                <h:outputText value="#{courseView.selectedCourse.category}"/>
 
                 <h:outputText value="Professor:"/>
-                <h:outputText value="#{courseSelection.selectedCourse.professor}"/>
+                <h:outputText value="#{courseView.selectedCourse.professor}"/>
 
             </p:panelGrid>
         </p:outputPanel>
         <f:facet name="footer">
             <h:panelGroup columns="2" columnClasses="button,button">
-                <p:commandButton value="Enroll" actionListener="#{enrollmentHandler.enroll(loginDto, courseSelection.selectedCourse)}" oncomplete="window.alert('Enrollment successful!');"/>
+                <p:commandButton value="Enroll" actionListener="#{enrollmentView.enroll(loginView.email, courseView.selectedCourse)}" />
                 <p:commandButton style="float:right" value="Cancel" onclick="PF('courseDialog').hide();" />
             </h:panelGroup>
         </f:facet>
