--- src/main/webapp/template/ui-template.xhtml
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-<html xmlns="http://www.w3.org/1999/xhtml"
-      xmlns:h="http://java.sun.com/jsf/html"
-      xmlns:ui="http://java.sun.com/jsf/facelets"
-      xmlns:p="http://primefaces.org/ui">
-
-<h:head>
-    <title>E-Learning <ui:insert name="pageTitle"/></title>
-    <h:outputStylesheet library="css" name="common.css"/>
-   <!-- <h:outputScript name="js/common.js"/>
-    <script>
-        $(document).on("pfAjaxError", function(event, xhr, options) {
-            Adamo.Utils.showErrorDialog(event, xhr, options);
-        });
-    </script>
-    <ui:insert name="additionalHeadResources"/> -->
-</h:head>
-<h:body>
-
-</h:body>
-</html>
